revert babel name refactor
