+++ flask_appbuilder/babel/manager.py
@@ -1,7 +1,7 @@
 import os
 
 from flask import has_request_context, request, session
-from flask_appbuilder.babels.views import LocaleView
+from flask_appbuilder.babel.views import LocaleView
 from flask_appbuilder.basemanager import BaseManager
 from flask_babel import Babel
 
@@ -20,9 +20,7 @@ class BabelManager(BaseManager):
         appbuilder_parent_dir = os.path.join(
             os.path.dirname(os.path.abspath(__file__)), os.pardir
         )
-        appbuilder_translations_path = os.path.join(
-            appbuilder_parent_dir, "translations"
-        )
+        appbuilder_translations_path = os.path.join(appbuilder_parent_dir, "translations")
         if "BABEL_TRANSLATION_DIRECTORIES" in app.config:
             current_translation_directories = app.config.get(
                 "BABEL_TRANSLATION_DIRECTORIES"
