+++ flask_appbuilder/babel/views.py
